Conflicts:
	App-Info.plist

+++ lib/ui/BUIconActionSheet.h
@@ -17,7 +17,7 @@ enum  {
 	BUIconActionSheetIconTypeCopy		= 995,
 
 };
-typedef int BUIconActionSheetIconType;
+typedef NSInteger BUIconActionSheetIconType;
