+++ lib/utilities/NSOrderedDictionary.m
@@ -289,7 +289,7 @@
         return [obj isEqual:key];
     }];
     NSInteger index = NSNotFound;
-    unsigned current_index = [idx1 firstIndex];
+    NSUInteger current_index = [idx1 firstIndex];
     while (current_index != NSNotFound && index == NSNotFound)
     {
         if ([idx2 containsIndex:current_index]) {
@@ -324,7 +324,7 @@
         return [obj isEqual:key];
     }];
     NSInteger index = NSNotFound;
-    unsigned current_index = [idx1 firstIndex];
+    NSInteger current_index = [idx1 firstIndex];
     while (current_index != NSNotFound && index == NSNotFound)
     {
         if ([idx2 containsIndex:current_index]) {
